pyproject.toml
[tool.poetry]
name = "gameLeaderboard"
version = "0.1.0"
description = "Game Leaderboard with Django."
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
drf-yasg = "^1.20.0"
django-redis = "^4.12.1"
django-filter = "^2.4.0"
orjson = "^3.4.6"
django-heroku = "^0.3.1"
django-silk = "^4.1.0"
psycopg2-binary = "^2.8.6"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
